Add old FHI-aims GNU recipe
This recipe is from 2018 and targets Singulary 2.6.  Despite the
version difference, it still mostly works, up until the build of
FHI-aims where it complains there's no scalapack.mpi target.

Before this can be addressed, though, the installation of GCC from
source needs to be removed, as this makes the build time obscenely
long (over an hour).