Add preliminary BerkeleyGW OneAPI build
This build suffers from the severe restriction that BerkeleyGW must
be launched from within the container; this is because the Intel
environment script cannot be run in sh shell, not allowing a "clean"
way to set environment variables using the script.  This functionally
makes it a Docker-esque container implemented in Singularity.

Additionally, this build exposes the large amount of copy-paste
inherent in segregating out code install among builds.